chore: DBTP-1596 Move platform_config_schema.py to providers
