+++ tests/platform_helper/utils/test_validation.py
@@ -8,9 +8,9 @@ from schema import SchemaError
 
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
-from dbt_platform_helper.utils.platform_config_schema import _is_integer_between
-from dbt_platform_helper.utils.platform_config_schema import _string_matching_regex
-from dbt_platform_helper.utils.platform_config_schema import _valid_s3_bucket_name
+from dbt_platform_helper.providers.platform_config_schema import _is_integer_between
+from dbt_platform_helper.providers.platform_config_schema import _string_matching_regex
+from dbt_platform_helper.providers.platform_config_schema import _valid_s3_bucket_name
 from dbt_platform_helper.utils.validation import _validate_extension_supported_versions
 from dbt_platform_helper.utils.validation import config_file_check
 from dbt_platform_helper.utils.validation import float_between_with_halfstep
